Update check_input_data
Expected data frame column names and their types and names are in a master data frame which is processed for each data frame and column with a shared code. This makes it easier to add tests for any new data frames and columns later.

Now, the function also returns the names of the columns that are missing in a df. Earlier the function only returned that some columns are missing but not which ones.